requirements.txt
azure-functions
azure-cosmos
openai
python-docx
reportlab
python-dotenv
azure-functions-durable